Add find event command
