+++ src/main/java/seedu/address/logic/commands/FindEventCommand.java
@@ -1,0 +1,60 @@
+package seedu.address.logic.commands;
+
+import static java.util.Objects.requireNonNull;
+
+import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.logic.Messages;
+import seedu.address.model.Model;
+import seedu.address.model.event.EventNameContainsKeywordsPredicate;
+import seedu.address.ui.CommandTabChange;
+
+/**
+ * Finds and lists all events in address book whose name contains any of the argument keywords.
+ * Keyword matching is case insensitive.
+ */
+public class FindEventCommand extends FindCommand {
+
+    public static final String COMMAND_FIELD = "event";
+
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds all events whose names contain any of "
+            + "the specified keywords (case-insensitive) and displays them as a list with index numbers.\n"
+            + "Parameters: KEYWORD [MORE_KEYWORDS]...\n"
+            + "Example: " + COMMAND_WORD + " alice bob charlie";
+
+    private final EventNameContainsKeywordsPredicate predicate;
+
+    public FindEventCommand(EventNameContainsKeywordsPredicate predicate) {
+        this.predicate = predicate;
+    }
+
+    @Override
+    public CommandResult execute(Model model) {
+        requireNonNull(model);
+        model.updateFilteredEventList(predicate);
+        return new CommandResult(
+                String.format(Messages.MESSAGE_EVENTS_LISTED_OVERVIEW, model.getFilteredEventList().size()),
+                false, false, CommandTabChange.EVENT);
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (other == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(other instanceof FindEventCommand)) {
+            return false;
+        }
+
+        FindEventCommand otherFindEventCommand = (FindEventCommand) other;
+        return predicate.equals(otherFindEventCommand.predicate);
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this)
+                .add("predicate", predicate)
+                .toString();
+    }
+}
